refactor: colonizethis_data wave 4 slice C victory config topic split (#4121)
Split ai_victory_config.dart into pace, observer, stalled-ow, declare-war,
offer-peace, colonial, and work topic libraries with a thin re-export façade.
Registry parity and lib file size gates remain green.
